refactor: easterEgg 디렉토리 구조 병합
