+++ src/main/java/everTale/everTale_be/domain/easterEgg/dto/easterEggVoice/request/EasterEggVoiceRegisterRequestDto.java
@@ -1,4 +1,4 @@
-package everTale.everTale_be.domain.easterEggVoice.dto.request;
+package everTale.everTale_be.domain.easterEgg.dto.easterEggVoice.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
