+++ src/main/java/everTale/everTale_be/domain/easterEgg/service/EasterEggLetterService.java
@@ -1,8 +1,8 @@
-package everTale.everTale_be.domain.easterEggLetter.service;
+package everTale.everTale_be.domain.easterEgg.service;
 
-import everTale.everTale_be.domain.easterEggLetter.dto.EasterEggLetterRequestDTO;
-import everTale.everTale_be.domain.easterEggLetter.dto.EasterEggLetterResponseDTO;
-import everTale.everTale_be.domain.easterEggLetter.entity.EasterEggLetter;
+import everTale.everTale_be.domain.easterEgg.dto.easterEggLetter.EasterEggLetterRequestDTO;
+import everTale.everTale_be.domain.easterEgg.dto.easterEggLetter.EasterEggLetterResponseDTO;
+import everTale.everTale_be.domain.easterEgg.entity.EasterEggLetter;
 import everTale.everTale_be.domain.profile.entity.Enum.ProfileType;
 import everTale.everTale_be.domain.profile.entity.Profile;
 import everTale.everTale_be.domain.profile.util.ProfileHelper;
